requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
amqp==2.4.2               # via kombu
appdirs==1.4.3            # via black
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, pytest
billiard==3.6.0.0         # via celery
black==19.3b0
celery[gevent,redis]==4.3.0
click==7.0                # via black
dj-database-url==0.5.0
django==2.2.1
entrypoints==0.3          # via flake8
flake8==3.7.7
gevent==1.4.0             # via celery
greenlet==0.4.15          # via gevent
isort==4.3.17
kombu==4.5.0              # via celery
mccabe==0.6.1             # via flake8
more-itertools==7.0.0     # via pytest
pluggy==0.9.0             # via pytest
psycopg2-binary==2.8.2
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pytest-django==3.4.8
pytest==4.4.1
python-dotenv==0.10.1
pytz==2019.1              # via celery, django
raven==6.10.0
redis==3.2.1              # via celery
shortuuid==0.5.0
six==1.12.0               # via pytest
sqlparse==0.3.0           # via django
toml==0.10.0              # via black
uwsgi==2.0.18
vine==1.3.0               # via amqp, celery
whitenoise==4.1.2
